# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "自控力API",
        "title": "Self_Control API",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "TAODEI",
            "email": "864978550@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "39.99.53.8:2333",
    "basePath": "/api/v1",
    "paths": {
        "/backdrop": {
            "get": {
                "description": "获取背景价格",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backdrop"
                ],
                "summary": "背景价格",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "根据背景id兑换背景",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backdrop"
                ],
                "summary": "兑换背景",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "backdrop_id",
                        "name": "backdrop_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BackdropID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"兑换成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "203": {
                        "description": "金币不足"
                    },
                    "400": {
                        "description": "{\"error_code\":\"20001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
        "/backdrops": {
            "get": {
                "description": "获取我的背景id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backdrop"
                ],
                "summary": "我的背景",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
        "/list/history": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "获取兑换排名历史",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "203": {
                        "description": "金币不足"
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
        "/list/user/{id}/{type}": {
            "get": {
                "description": "根据 type 和 id 获取用户排名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "用户排名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
        "/list/{type}": {
            "put": {
                "description": "需要前进的排名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "兑换排名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ranking",
                        "name": "ranking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Ranking"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"兑换成功\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "203": {
                        "description": "金币不足"
                    },
                    "400": {
                        "description": "{\"error_code\":\"20001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "401": {
                        "description": "{\"error_code\":\"10001\", \"message\":\"Token Invalid.\"} 身份认证失败 重新登录",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
        "/lists/{type}": {
            "get": {
                "description": "url最后面+week或month查询数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "排行榜数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\":\"获取前五用户\"}",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "203": {
                        "description": "未检索到该时间段的打卡信息"
                    },
                    "400": {
                        "description": "{\"error_code\":\"20001\", \"message\":\"Fail.\"} or {\"error_code\":\"00002\", \"message\":\"Lack Param Or Param Not Satisfiable.\"}",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    },
                    "500": {
                        "description": "{\"error_code\":\"30001\", \"message\":\"Fail.\"} 失败",
                        "schema": {
                            "$ref": "#/definitions/error.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "error.Error": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.BackdropID": {
            "type": "object",
            "properties": {
                "backdrop_id": {
                    "type": "integer"
                }
            }
        },
        "model.Ranking": {
            "type": "object",
            "properties": {
                "ranking": {
                    "type": "integer"
                }
            }
        },
...
    }
}